chore(wizard-refactor): finalize — dispatcher → index.ts, monolith → barrel shim
lib/chat/wizard-tool-executor.ts is now a 6-line re-export shim. The
83-line dispatcher lives at lib/chat/wizard-tool-executor/index.ts.
External callers (chat/route.ts, ConversationalWizard.tsx, the four
wizard test files) keep working without import-line churn.

Final tree:
  wizard-tool-executor.ts                 (6 lines — re-export shim)
  wizard-tool-executor/
    index.ts                              (83 lines — dispatcher)
    _shared/                              (5 helpers, 198 lines total)
    resolvers/                            (4 files, 323 lines total)
    tools/                                (10 files; create_course=1361,
                                           update_setup=372, rest <235 each)

Acceptance criteria from HANDOFF-wizard-refactor-split.md:
  [x] All 10 case bodies → tools/<name>.ts
  [x] All 4 resolvers   → resolvers/<name>.ts
  [x] All 5 helpers     → _shared/
  [x] No new file >1000 lines except create_course (tracked as
      follow-on HF-Q per handoff §Open question)
  [x] Dispatcher 11/11 + lint 17/17 + sibling suites 33/33 green
      (61/61 total wizard tests)
  [x] Imports use canonical @/lib/ alias (one relative import
      promoted in create_course)
